sol, expired, SQL fixed :)
auto remove from stock to sold, fix expired and fix and connect table to others SQL

+++ purchase.php
@@ -45,11 +45,15 @@ session_start();
       header("location: purchase.php");
    }
 
-   #-------- delete db from purchase---------------
+   #-------- delete db from purchase and put it in trash table---------------
    if(isset($_GET['id'])){
       $id = $_GET['id'];
-      $sqls= "DELETE from purchase where id='$id'";
-      mysqli_query($con,$sqls);
+      $sql1 = "INSERT INTO ptrash SELECT * FROM purchase WHERE id='$id'";
+      mysqli_query($con,$sql1);
+      if(true){
+         $sqls= "DELETE from purchase where id='$id'";
+         mysqli_query($con,$sqls);
+      }
       header("location: purchase.php");
    }
 
@@ -81,6 +85,7 @@ session_start();
                <li><a  href="note.php"><img class="fas"src="icons/contacts.svg">Notes</a></li>
                <li class="active"><a href="purchase.php"><img class="fas"src="icons/buy.svg">Purchase</a></li>
                <li><a href="sells.php"><img class="fas"src="icons/sells.svg">Sells</a></li>
+               <li><a href="sold.php"><img class="fas"src="icons/sells.svg">Sold</a></li>
                <li><a href="stock.php"><img class="fas"src="icons/stock.svg">Stock</a></li>
                <li><a href="expired.php"><img class="fas"src="icons/expired.svg">Expired Products</a></li>
                <li><a href="logout.php"><img class="fas"src="icons/logout.svg">Logout</a></li>
@@ -113,7 +118,7 @@ session_start();
                   <input type="number" name="price" id="price" placeholder="price" min="1"><br>
                   
                   <button name="add">ADD</button>
-                  <button name="del">Delete</button>
+                  <!-- <button name="del">Delete</button> -->
                </div>
             </aside>
             <!--table-->
@@ -146,5 +151,8 @@ session_start();
             </main>
          </form>
       </div>
+   </div>
+<script>
+</script>
 </body>
 </html>
